Fix signedness
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@891 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueSettings.cpp
@@ -135,8 +135,8 @@ void GOrgueSettings::Load()
 	m_DefaultAudioDevice = m_Config.Read(wxT("Devices/DefaultSound"), wxEmptyString);
 
 	m_OrganMidiEvents.clear();
-	long count = m_Config.Read(wxT("OrganMIDI/Count"), 0L);
-	for (long i = 0; i < count; i++)
+	unsigned count = m_Config.Read(wxT("OrganMIDI/Count"), 0L);
+	for (unsigned i = 0; i < count; i++)
 	{
 		wxString itemstr = wxT("OrganMIDI/Organ") + wxString::Format(wxT("%ld"), i);
 		long j = m_Config.Read(itemstr + wxT(".midi"), 0L);
